requirements.txt
contextlib2==0.5.5
Django==1.10.8
django-nose==1.4.5
djangorestframework==3.5.3
django-health-check==3.7.0
docutils==0.14
furl==1.2.1
mysqlclient==1.3.13
nose==1.3.7
py-auth0-jwt==0.3.2
python-dateutil==2.6.1
raven==6.9.0
requests==2.19.1
six==1.10.0
